Add empty blog posts controller
We add an empty controller that inherits from our app’s base controller class. We could have gotten past this error message by creating a class that didn’t inherit from anything, but in this case we’re so sure we’ll inherit from the base controller class that we can go ahead and do it.

Red: The action 'new' could not be found for BlogPostsController

The acceptance test can now find the controller, but not a "new" action on it.