Add create method to blog post controller
We add the `create` method to the blog post controller directly, then rerun the acceptance test.

Outer red: Unable to find xpath "/html"

The error we get is a bit obscure, but effectively it means the acceptance test can't even find an HTML document returned to search within. Unlike with the `new` action, with the `create` action, if no explicit render or redirect happens, Rails 5 does not return an error that a template is missing; it just returns empty content.